refactor(console): remove -ng suffix
