+++ gravitee-apim-console-webui/src/management/application/applications.module.ts
@@ -34,7 +34,7 @@ import { ApplicationNavigationModule } from './application-navigation/applicatio
 import { ApplicationNotificationSettingsModule } from './details/notifications/notification-settings/application-notification-settings.module';
 import { ApplicationMetadataModule } from './details/metadata/application-metadata.module';
 import { ApplicationGeneralUserGroupModule } from './details/user-group-access/application-general-user-group.module';
-import { ApplicationGeneralNgModule } from './details/general/general-ng/application-general-ng.module';
+import { ApplicationGeneralModule } from './details/general/application-general.module';
 import { ApplicationSubscriptionsComponent } from './details/subscriptions/application-subscriptions.component';
 import { ApplicationSubscriptionComponent } from './details/subscriptions/application-subscription.component';
 import { ApplicationAnalyticsComponent } from './details/analytics/application-analytics.component';
@@ -73,7 +73,7 @@ import { GioPermissionModule } from '../../shared/components/gio-permission/gio-
     ApplicationNotificationSettingsModule,
     ApplicationMetadataModule,
     ApplicationGeneralUserGroupModule,
-    ApplicationGeneralNgModule,
+    ApplicationGeneralModule,
   ],
   declarations: [
     EnvApplicationListComponent,
