+++ gravitee-apim-console-webui/src/management/application/details/general/application-general.component.html
